revert a small change made in range calculation in commit 8431a765ea07631b36eb0befa3b6cd5d8a7c377a from 28 Nov 2020

+++ src/Common/Community.cpp
@@ -1631,10 +1631,10 @@ outrange << "\t" << s.occupied << "\t" << occsuit;
 // RANGE MINIMA AND MAXIMA NEED TO BECOME A PROPERTY OF THE SPECIES
 if (s.ninds > 0) {
 	landOrigin origin = pLandscape->getOrigin();
-	outrange << "\t" << s.minX * ppLand.resol + origin.minEast
-		<< "\t" << s.maxX+1 * ppLand.resol + origin.minEast
-		<< "\t" << s.minY * ppLand.resol + origin.minNorth
-		<< "\t" << s.maxY+1 * ppLand.resol + origin.minNorth;
+	outrange << "\t" << (float)s.minX * (float)ppLand.resol + origin.minEast
+		<< "\t" << (float)(s.maxX+1) * (float)ppLand.resol + origin.minEast
+		<< "\t" << (float)s.minY * (float)ppLand.resol + origin.minNorth
+		<< "\t" << (float)(s.maxY+1) * (float)ppLand.resol + origin.minNorth;
 }
 else
 	outrange <<"\t0\t0\t0\t0";
